Extract stats playback parsing modules
