+++ src/collector/stats/playback.rs
@@ -1,4 +1,3 @@
-use boxcars::{Ps4Id, PsyNetId, RemoteId, SwitchId};
 use serde::de::DeserializeOwned;
 use serde::Serialize;
 use serde_json::{Map, Value};
@@ -6,6 +5,16 @@ use serde_json::{Map, Value};
 use crate::*;
 
 use super::types::serialize_to_json_value;
+
+#[path = "playback_event_parsers.rs"]
+mod playback_event_parsers;
+#[path = "playback_json.rs"]
+mod playback_json;
+#[path = "playback_mechanics.rs"]
+mod playback_mechanics;
+use playback_event_parsers::*;
+use playback_json::*;
+use playback_mechanics::*;
 
 #[derive(Debug, Clone, PartialEq, Serialize)]
 pub struct CapturedStatsFrame<Modules> {
@@ -2007,1538 +2016,4 @@ impl CapturedStatsData<ReplayStatsFrame> {
     pub fn into_replay_stats_timeline(self) -> SubtrActorResult<ReplayStatsTimeline> {
         self.into_legacy_replay_stats_timeline()
     }
-}
-
-fn player_stats_value_for_key<'a>(
-    module: Option<&'a Value>,
-    player_key: &str,
-) -> SubtrActorResult<Option<&'a Value>> {
-    let Some(entries) = module
-        .and_then(Value::as_object)
-        .and_then(|module| module.get("player_stats"))
-        .and_then(Value::as_array)
-    else {
-        return Ok(None);
-    };
-
-    for entry in entries {
-        let Some(entry_object) = entry.as_object() else {
-            continue;
-        };
-        let Some(player_id) = entry_object.get("player_id") else {
-            continue;
-        };
-        let Some(player_stats) = entry_object.get("stats") else {
-            continue;
-        };
-        if player_id_key(player_id)? == player_key {
-            return Ok(Some(player_stats));
-        }
-    }
-
-    Ok(None)
-}
-
-fn player_info_key(player: &PlayerInfo) -> SubtrActorResult<String> {
-    player_id_key(&serialize_to_json_value(&player.remote_id)?)
-}
-
-fn player_id_key(player_id: &Value) -> SubtrActorResult<String> {
-    serde_json::to_string(player_id).map_err(|error| {
-        SubtrActorError::new(SubtrActorErrorVariant::StatsSerializationError(
-            error.to_string(),
-        ))
-    })
-}
-
-fn default_json_value<T>() -> Value
-where
-    T: Default + Serialize,
-{
-    serde_json::to_value(T::default()).expect("default stats should serialize to json")
-}
-
-fn decode_json_value<T>(value: Value) -> SubtrActorResult<T>
-where
-    T: DeserializeOwned,
-{
-    serde_json::from_value(value).map_err(|error| {
-        SubtrActorError::new(SubtrActorErrorVariant::StatsSerializationError(
-            error.to_string(),
-        ))
-    })
-}
-
-fn decode_core_player_stats_value(mut value: Value) -> SubtrActorResult<CorePlayerStats> {
-    normalize_core_player_stats_snapshot(&mut value)?;
-    decode_json_value(value)
-}
-
-fn normalize_core_player_stats_snapshot(value: &mut Value) -> SubtrActorResult<()> {
-    let Some(object) = value.as_object_mut() else {
-        return Ok(());
-    };
-
-    insert_cumulative_from_average(
-        object,
-        "cumulative_boost_on_goals_against",
-        "average_boost_on_goals_against",
-        "goal_against_boost_sample_count",
-    )?;
-    insert_cumulative_from_average(
-        object,
-        "cumulative_average_boost_in_goal_against_leadup",
-        "average_boost_in_goal_against_leadup",
-        "goal_against_boost_leadup_sample_count",
-    )?;
-    insert_cumulative_from_average(
-        object,
-        "cumulative_min_boost_in_goal_against_leadup",
-        "average_min_boost_in_goal_against_leadup",
-        "goal_against_boost_leadup_sample_count",
-    )?;
-    insert_cumulative_from_average(
-        object,
-        "cumulative_goal_against_position_x",
-        "average_goal_against_position_x",
-        "goal_against_position_sample_count",
-    )?;
-    insert_cumulative_from_average(
-        object,
-        "cumulative_goal_against_position_y",
-        "average_goal_against_position_y",
-        "goal_against_position_sample_count",
-    )?;
-    insert_cumulative_from_average(
-        object,
-        "cumulative_goal_against_position_z",
-        "average_goal_against_position_z",
-        "goal_against_position_sample_count",
-    )?;
-    insert_cumulative_from_average(
-        object,
-        "cumulative_scoring_goal_last_touch_position_x",
-        "average_scoring_goal_last_touch_position_x",
-        "scoring_goal_last_touch_position_sample_count",
-    )?;
-    insert_cumulative_from_average(
-        object,
-        "cumulative_scoring_goal_last_touch_position_y",
-        "average_scoring_goal_last_touch_position_y",
-        "scoring_goal_last_touch_position_sample_count",
-    )?;
-    insert_cumulative_from_average(
-        object,
-        "cumulative_scoring_goal_last_touch_position_z",
-        "average_scoring_goal_last_touch_position_z",
-        "scoring_goal_last_touch_position_sample_count",
-    )?;
-    insert_cumulative_from_average(
-        object,
-        "cumulative_goal_ball_air_time",
-        "average_goal_ball_air_time",
-        "goal_ball_air_time_sample_count",
-    )?;
-
-    if let Value::Object(defaults) = default_json_value::<CorePlayerStats>() {
-        for (field, default_value) in defaults {
-            object.entry(field).or_insert(default_value);
-        }
-    }
-
-    Ok(())
-}
-
-fn insert_cumulative_from_average(
-    object: &mut Map<String, Value>,
-    cumulative_field: &str,
-    average_field: &str,
-    sample_count_field: &str,
-) -> SubtrActorResult<()> {
-    if object.contains_key(cumulative_field) {
-        return Ok(());
-    }
-
-    let average = object
-        .get(average_field)
-        .and_then(Value::as_f64)
-        .unwrap_or(0.0) as f32;
-    let sample_count = object
-        .get(sample_count_field)
-        .and_then(Value::as_u64)
-        .unwrap_or(0) as f32;
-    object.insert(
-        cumulative_field.to_owned(),
-        serialize_to_json_value(&(average * sample_count))?,
-    );
-
-    Ok(())
-}
-
-fn parse_timeline_event(value: &Value) -> SubtrActorResult<TimelineEvent> {
-    let object = json_object(value, "timeline event")?;
-    Ok(TimelineEvent {
-        time: json_required_f32(object, "time")?,
-        frame: json_optional_usize(object.get("frame"))?,
-        kind: decode_json_value(json_required_value(object, "kind")?.clone())?,
-        player_id: json_optional_remote_id(object.get("player_id"))?,
-        player_position: json_optional_vec3(object.get("player_position"))?,
-        is_team_0: json_optional_bool(object.get("is_team_0")),
-    })
-}
-
-fn moment_mechanic_event(
-    kind: &str,
-    index: usize,
-    frame: usize,
-    time: f32,
-    player_id: PlayerId,
-    player_position: Option<[f32; 3]>,
-    is_team_0: bool,
-) -> StatsTimelineTagEvent {
-    StatsTimelineTagEvent {
-        id: format!("{kind}:{frame}:{index}"),
-        kind: kind.to_owned(),
-        player_id,
-        player_position,
-        is_team_0,
-        timing: StatsEventTiming::Moment { frame, time },
-        properties: Vec::new(),
-    }
-}
-
-#[allow(clippy::too_many_arguments)]
-fn span_mechanic_event(
-    kind: &str,
-    index: usize,
-    start_frame: usize,
-    end_frame: usize,
-    start_time: f32,
-    end_time: f32,
-    player_id: PlayerId,
-    player_position: Option<[f32; 3]>,
-    is_team_0: bool,
-) -> StatsTimelineTagEvent {
-    StatsTimelineTagEvent {
-        id: format!("{kind}:{start_frame}:{end_frame}:{index}"),
-        kind: kind.to_owned(),
-        player_id,
-        player_position,
-        is_team_0,
-        timing: StatsEventTiming::Span {
-            start_frame,
-            end_frame,
-            start_time,
-            end_time,
-        },
-        properties: Vec::new(),
-    }
-}
-
-fn mechanic_event_start_time(event: &StatsTimelineTagEvent) -> f32 {
-    match event.timing {
-        StatsEventTiming::Moment { time, .. } => time,
-        StatsEventTiming::Span { start_time, .. } => start_time,
-    }
-}
-
-fn mechanic_event_text_property(key: &str, value: &str) -> StatsEventProperty {
-    StatsEventProperty {
-        key: key.to_owned(),
-        value: StatsEventPropertyValue::Text(value.to_owned()),
-    }
-}
-
-fn mechanic_event_unsigned_property(key: &str, value: u32) -> StatsEventProperty {
-    StatsEventProperty {
-        key: key.to_owned(),
-        value: StatsEventPropertyValue::Unsigned(value),
-    }
-}
-
-fn ball_carry_mechanic_event_properties(
-    object: &serde_json::Map<String, Value>,
-) -> Vec<StatsEventProperty> {
-    let mut properties = Vec::new();
-    if let Some(origin) = object.get("air_dribble_origin").and_then(Value::as_str) {
-        properties.push(mechanic_event_text_property("origin", origin));
-    }
-    if let Some(touch_count) = object.get("touch_count").and_then(Value::as_u64) {
-        properties.push(mechanic_event_unsigned_property(
-            "touch_count",
-            touch_count as u32,
-        ));
-    }
-    properties
-}
-
-fn parse_ball_carry_mechanic_event(
-    value: &Value,
-    index: usize,
-) -> SubtrActorResult<StatsTimelineTagEvent> {
-    let object = json_object(value, "ball carry mechanic event")?;
-    let serialized_kind = json_required_str(object, "kind")?;
-    let kind = match serialized_kind {
-        "carry" => "ball_carry",
-        "air_dribble" => "air_dribble",
-        other => other,
-    };
-    let mut mechanic_event = span_mechanic_event(
-        kind,
-        index,
-        json_required_usize(object, "start_frame")?,
-        json_required_usize(object, "end_frame")?,
-        json_required_f32(object, "start_time")?,
-        json_required_f32(object, "end_time")?,
-        json_required_remote_id(object, "player_id")?,
-        Some(json_required_vec3(object, "end_position")?),
-        json_required_bool(object, "is_team_0")?,
-    );
-    if kind == "air_dribble" {
-        mechanic_event.properties = ball_carry_mechanic_event_properties(object);
-    }
-    Ok(mechanic_event)
-}
-
-fn parse_dodge_reset_mechanic_event(
-    value: &Value,
-    index: usize,
-) -> SubtrActorResult<StatsTimelineTagEvent> {
-    let object = json_object(value, "dodge reset mechanic event")?;
-    Ok(moment_mechanic_event(
-        "flip_reset",
-        index,
-        json_required_usize(object, "frame")?,
-        json_required_f32(object, "time")?,
-        json_required_remote_id(object, "player")?,
-        json_optional_vec3(object.get("player_position"))?,
-        json_required_bool(object, "is_team_0")?,
-    ))
-}
-
-fn parse_dodge_reset_event(value: &Value) -> SubtrActorResult<DodgeResetEvent> {
-    let object = json_object(value, "dodge reset event")?;
-    Ok(DodgeResetEvent {
-        time: json_required_f32(object, "time")?,
-        frame: json_required_usize(object, "frame")?,
-        player: json_required_remote_id(object, "player")?,
-        is_team_0: json_required_bool(object, "is_team_0")?,
-        player_position: json_optional_vec3(object.get("player_position"))?,
-        counter_value: json_required_i32(object, "counter_value")?,
-        on_ball: json_required_bool(object, "on_ball")?,
-    })
-}
-
-fn parse_powerslide_event(value: &Value) -> SubtrActorResult<PowerslideEvent> {
-    let object = json_object(value, "powerslide event")?;
-    Ok(PowerslideEvent {
-        time: json_required_f32(object, "time")?,
-        frame: json_required_usize(object, "frame")?,
-        player: json_required_remote_id(object, "player")?,
-        player_position: json_optional_vec3(object.get("player_position"))?,
-        is_team_0: json_required_bool(object, "is_team_0")?,
-        active: json_required_bool(object, "active")?,
-    })
-}
-
-fn parse_core_player_stats_event(value: &Value) -> SubtrActorResult<CorePlayerStatsEvent> {
-    let object = json_object(value, "core player stats event")?;
-    Ok(CorePlayerStatsEvent {
-        time: json_required_f32(object, "time")?,
-        frame: json_required_usize(object, "frame")?,
-        player: json_required_remote_id(object, "player")?,
-        player_position: json_optional_vec3(object.get("player_position"))?,
-        is_team_0: json_required_bool(object, "is_team_0")?,
-        delta: decode_json_value(json_required_value(object, "delta")?.clone())?,
-    })
-}
-
-fn parse_core_team_stats_event(value: &Value) -> SubtrActorResult<CoreTeamStatsEvent> {
-    let object = json_object(value, "core team stats event")?;
-    Ok(CoreTeamStatsEvent {
-        time: json_required_f32(object, "time")?,
-        frame: json_required_usize(object, "frame")?,
-        is_team_0: json_required_bool(object, "is_team_0")?,
-        delta: decode_json_value(json_required_value(object, "delta")?.clone())?,
-    })
-}
-
-fn parse_possession_event(value: &Value) -> SubtrActorResult<PossessionEvent> {
-    let object = json_object(value, "possession event")?;
-    Ok(PossessionEvent {
-        time: json_required_f32(object, "time")?,
-        frame: json_required_usize(object, "frame")?,
-        active: json_required_bool(object, "active")?,
-        possession_state: json_required_str(object, "possession_state")?.to_owned(),
-        field_third: match object.get("field_third") {
-            None | Some(Value::Null) => None,
-            Some(_) => Some(json_required_str(object, "field_third")?.to_owned()),
-        },
-    })
-}
-
-fn parse_pressure_event(value: &Value) -> SubtrActorResult<PressureEvent> {
-    let object = json_object(value, "pressure event")?;
-    Ok(PressureEvent {
-        time: json_required_f32(object, "time")?,
-        frame: json_required_usize(object, "frame")?,
-        active: json_required_bool(object, "active")?,
-        field_half: json_required_str(object, "field_half")?.to_owned(),
-    })
-}
-
-fn parse_territorial_pressure_event(value: &Value) -> SubtrActorResult<TerritorialPressureEvent> {
-    decode_json_value(value.clone())
-}
-
-fn parse_movement_event(value: &Value) -> SubtrActorResult<MovementEvent> {
-    let object = json_object(value, "movement event")?;
-    Ok(MovementEvent {
-        time: json_required_f32(object, "time")?,
-        frame: json_required_usize(object, "frame")?,
-        player: json_required_remote_id(object, "player")?,
-        player_position: json_optional_vec3(object.get("player_position"))?,
-        is_team_0: json_required_bool(object, "is_team_0")?,
-        dt: json_required_f32(object, "dt")?,
-        speed: json_required_f32(object, "speed")?,
-        distance: json_required_f32(object, "distance")?,
-        speed_band: json_required_str(object, "speed_band")?.to_owned(),
-        height_band: json_required_str(object, "height_band")?.to_owned(),
-    })
-}
-
-fn parse_positioning_event(value: &Value) -> SubtrActorResult<PositioningEvent> {
-    let object = json_object(value, "positioning event")?;
-    Ok(PositioningEvent {
-        time: json_required_f32(object, "time")?,
-        frame: json_required_usize(object, "frame")?,
-        player: json_required_remote_id(object, "player")?,
-        player_position: json_optional_vec3(object.get("player_position"))?,
-        is_team_0: json_required_bool(object, "is_team_0")?,
-        active_game_time: json_required_f32(object, "active_game_time")?,
-        tracked_time: json_required_f32(object, "tracked_time")?,
-        sum_distance_to_teammates: json_required_f32(object, "sum_distance_to_teammates")?,
-        sum_distance_to_ball: json_required_f32(object, "sum_distance_to_ball")?,
-        sum_distance_to_ball_has_possession: json_required_f32(
-            object,
-            "sum_distance_to_ball_has_possession",
-        )?,
-        time_has_possession: json_required_f32(object, "time_has_possession")?,
-        sum_distance_to_ball_no_possession: json_required_f32(
-            object,
-            "sum_distance_to_ball_no_possession",
-        )?,
-        time_no_possession: json_required_f32(object, "time_no_possession")?,
-        time_demolished: json_required_f32(object, "time_demolished")?,
-        time_no_teammates: json_required_f32(object, "time_no_teammates")?,
-        time_most_back: json_required_f32(object, "time_most_back")?,
-        time_most_forward: json_required_f32(object, "time_most_forward")?,
-        time_mid_role: json_required_f32(object, "time_mid_role")?,
-        time_other_role: json_required_f32(object, "time_other_role")?,
-        time_defensive_zone: json_required_f32(object, "time_defensive_third")?,
-        time_neutral_zone: json_required_f32(object, "time_neutral_third")?,
-        time_offensive_zone: json_required_f32(object, "time_offensive_third")?,
-        time_defensive_half: json_required_f32(object, "time_defensive_half")?,
-        time_offensive_half: json_required_f32(object, "time_offensive_half")?,
-        time_closest_to_ball: json_required_f32(object, "time_closest_to_ball")?,
-        time_farthest_from_ball: json_required_f32(object, "time_farthest_from_ball")?,
-        time_behind_ball: json_required_f32(object, "time_behind_ball")?,
-        time_level_with_ball: json_required_f32(object, "time_level_with_ball")?,
-        time_in_front_of_ball: json_required_f32(object, "time_in_front_of_ball")?,
-        times_caught_ahead_of_play_on_conceded_goals: json_required_usize(
-            object,
-            "times_caught_ahead_of_play_on_conceded_goals",
-        )? as u32,
-    })
-}
-
-fn parse_rotation_player_event(value: &Value) -> SubtrActorResult<RotationPlayerEvent> {
-    let object = json_object(value, "rotation player event")?;
-    Ok(RotationPlayerEvent {
-        time: json_required_f32(object, "time")?,
-        frame: json_required_usize(object, "frame")?,
-        player: json_required_remote_id(object, "player")?,
-        player_position: json_optional_vec3(object.get("player_position"))?,
-        is_team_0: json_required_bool(object, "is_team_0")?,
-        active: json_required_bool(object, "active")?,
-        became_first_man_count: json_required_usize(object, "became_first_man_count")? as u32,
-        lost_first_man_count: json_required_usize(object, "lost_first_man_count")? as u32,
-        current_role_state: decode_json_value(
-            json_required_value(object, "current_role_state")?.clone(),
-        )?,
-        current_depth_state: decode_json_value(
-            json_required_value(object, "current_depth_state")?.clone(),
-        )?,
-    })
-}
-
-fn parse_rotation_team_event(value: &Value) -> SubtrActorResult<RotationTeamEvent> {
-    let object = json_object(value, "rotation team event")?;
-    Ok(RotationTeamEvent {
-        time: json_required_f32(object, "time")?,
-        frame: json_required_usize(object, "frame")?,
-        is_team_0: json_required_bool(object, "is_team_0")?,
-        first_man_changes_for_team: json_required_usize(object, "first_man_changes_for_team")?
-            as u32,
-        rotation_count: json_required_usize(object, "rotation_count")? as u32,
-    })
-}
-
-fn parse_touch_stats_event(value: &Value) -> SubtrActorResult<TouchStatsEvent> {
-    let object = json_object(value, "touch stats event")?;
-    let time = json_required_f32(object, "time")?;
-    let frame = json_required_usize(object, "frame")?;
-    Ok(TouchStatsEvent {
-        time,
-        frame,
-        sample_time: json_optional_f32(object.get("sample_time"))?.unwrap_or(time),
-        sample_frame: json_optional_usize(object.get("sample_frame"))?.unwrap_or(frame),
-        player: json_required_remote_id(object, "player")?,
-        player_position: json_optional_vec3(object.get("player_position"))?,
-        is_team_0: json_required_bool(object, "is_team_0")?,
-        kind: json_required_str(object, "kind")?.to_owned(),
-        height_band: json_required_str(object, "height_band")?.to_owned(),
-        surface: json_required_str(object, "surface")?.to_owned(),
-        dodge_state: json_required_str(object, "dodge_state")?.to_owned(),
-        ball_speed_change: json_required_f32(object, "ball_speed_change")?,
-    })
-}
-
-fn parse_touch_ball_movement_event(value: &Value) -> SubtrActorResult<TouchBallMovementEvent> {
-    let object = json_object(value, "touch ball movement event")?;
-    Ok(TouchBallMovementEvent {
-        time: json_required_f32(object, "time")?,
-        frame: json_required_usize(object, "frame")?,
-        player: json_required_remote_id(object, "player")?,
-        player_position: json_optional_vec3(object.get("player_position"))?,
-        is_team_0: json_required_bool(object, "is_team_0")?,
-        travel_distance: json_required_f32(object, "travel_distance")?,
-        advance_distance: json_required_f32(object, "advance_distance")?,
-        retreat_distance: json_required_f32(object, "retreat_distance")?,
-    })
-}
-
-fn parse_touch_last_touch_event(value: &Value) -> SubtrActorResult<TouchLastTouchEvent> {
-    let object = json_object(value, "touch last-touch event")?;
-    let time = json_required_f32(object, "time")?;
-    let frame = json_required_usize(object, "frame")?;
-    Ok(TouchLastTouchEvent {
-        time,
-        frame,
-        sample_time: json_optional_f32(object.get("sample_time"))?.unwrap_or(time),
-        sample_frame: json_optional_usize(object.get("sample_frame"))?.unwrap_or(frame),
-        is_team_0: json_required_bool(object, "is_team_0")?,
-        player: json_optional_remote_id(object.get("player"))?,
-        player_position: json_optional_vec3(object.get("player_position"))?,
-    })
-}
-
-fn parse_flick_mechanic_event(
-    value: &Value,
-    index: usize,
-) -> SubtrActorResult<StatsTimelineTagEvent> {
-    let object = json_object(value, "flick mechanic event")?;
-    Ok(span_mechanic_event(
-        "flick",
-        index,
-        json_required_usize(object, "setup_start_frame")?,
-        json_required_usize(object, "frame")?,
-        json_required_f32(object, "setup_start_time")?,
-        json_required_f32(object, "time")?,
-        json_required_remote_id(object, "player")?,
-        json_optional_vec3(object.get("player_position"))?,
-        json_required_bool(object, "is_team_0")?,
-    ))
-}
-
-fn parse_flick_event(value: &Value) -> SubtrActorResult<FlickEvent> {
-    let object = json_object(value, "flick event")?;
-    let time = json_required_f32(object, "time")?;
-    let frame = json_required_usize(object, "frame")?;
-    Ok(FlickEvent {
-        time,
-        frame,
-        sample_time: json_optional_f32(object.get("sample_time"))?.unwrap_or(time),
-        sample_frame: json_optional_usize(object.get("sample_frame"))?.unwrap_or(frame),
-        player: json_required_remote_id(object, "player")?,
-        player_position: json_optional_vec3(object.get("player_position"))?,
-        is_team_0: json_required_bool(object, "is_team_0")?,
-        dodge_time: json_required_f32(object, "dodge_time")?,
-        dodge_frame: json_required_usize(object, "dodge_frame")?,
-        time_since_dodge: json_required_f32(object, "time_since_dodge")?,
-        setup_start_time: json_required_f32(object, "setup_start_time")?,
-        setup_start_frame: json_required_usize(object, "setup_start_frame")?,
-        setup_duration: json_required_f32(object, "setup_duration")?,
-        setup_touch_count: json_required_usize(object, "setup_touch_count")? as u32,
-        average_horizontal_gap: json_required_f32(object, "average_horizontal_gap")?,
-        average_vertical_gap: json_required_f32(object, "average_vertical_gap")?,
-        ball_speed_change: json_required_f32(object, "ball_speed_change")?,
-        ball_impulse: json_required_vec3(object, "ball_impulse")?,
-        impulse_away_alignment: json_required_f32(object, "impulse_away_alignment")?,
-        vertical_impulse: json_required_f32(object, "vertical_impulse")?,
-        confidence: json_required_f32(object, "confidence")?,
-    })
-}
-
-fn parse_musty_flick_mechanic_event(
-    value: &Value,
-    index: usize,
-) -> SubtrActorResult<StatsTimelineTagEvent> {
-    let object = json_object(value, "musty flick mechanic event")?;
-    Ok(span_mechanic_event(
-        "musty_flick",
-        index,
-        json_required_usize(object, "dodge_frame")?,
-        json_required_usize(object, "frame")?,
-        json_required_f32(object, "dodge_time")?,
-        json_required_f32(object, "time")?,
-        json_required_remote_id(object, "player")?,
-        json_optional_vec3(object.get("player_position"))?,
-        json_required_bool(object, "is_team_0")?,
-    ))
-}
-
-fn parse_musty_flick_event(value: &Value) -> SubtrActorResult<MustyFlickEvent> {
-    let object = json_object(value, "musty flick event")?;
-    let time = json_required_f32(object, "time")?;
-    let frame = json_required_usize(object, "frame")?;
-    Ok(MustyFlickEvent {
-        time,
-        frame,
-        sample_time: json_optional_f32(object.get("sample_time"))?.unwrap_or(time),
-        sample_frame: json_optional_usize(object.get("sample_frame"))?.unwrap_or(frame),
-        player: json_required_remote_id(object, "player")?,
-        player_position: json_optional_vec3(object.get("player_position"))?,
-        is_team_0: json_required_bool(object, "is_team_0")?,
-        aerial: json_required_bool(object, "aerial")?,
-        dodge_time: json_required_f32(object, "dodge_time")?,
-        dodge_frame: json_required_usize(object, "dodge_frame")?,
-        time_since_dodge: json_required_f32(object, "time_since_dodge")?,
-        confidence: json_required_f32(object, "confidence")?,
-        local_ball_position: json_required_vec3(object, "local_ball_position")?,
-        rear_alignment: json_required_f32(object, "rear_alignment")?,
-        top_alignment: json_required_f32(object, "top_alignment")?,
-        forward_approach_speed: json_required_f32(object, "forward_approach_speed")?,
-        pitch_rate: json_required_f32(object, "pitch_rate")?,
-        ball_speed_change: json_required_f32(object, "ball_speed_change")?,
-    })
-}
-
-fn parse_goal_context_event(value: &Value) -> SubtrActorResult<GoalContextEvent> {
-    let object = json_object(value, "goal context event")?;
-    Ok(GoalContextEvent {
-        time: json_required_f32(object, "time")?,
-        frame: json_required_usize(object, "frame")?,
-        scoring_team_is_team_0: json_required_bool(object, "scoring_team_is_team_0")?,
-        scorer: json_optional_remote_id(object.get("scorer"))?,
-        scoring_team_most_back_player: json_optional_remote_id(
-            object.get("scoring_team_most_back_player"),
-        )?,
-        defending_team_most_back_player: json_optional_remote_id(
-            object.get("defending_team_most_back_player"),
-        )?,
-        ball_position: json_optional_goal_context_position(object.get("ball_position"))?,
-        ball_speed_at_goal: json_optional_f32(object.get("ball_speed_at_goal"))?,
-        ball_air_time_before_goal: json_optional_f32(object.get("ball_air_time_before_goal"))?,
-        goal_buildup: object
-            .get("goal_buildup")
-            .map(|value| decode_json_value(value.clone()))
-            .transpose()?
-            .unwrap_or_default(),
-        scorer_last_touch: match object.get("scorer_last_touch") {
-            None | Some(Value::Null) => None,
-            Some(value) => Some(parse_goal_touch_context(value)?),
-        },
-        players: json_required_array(object, "players")?
-            .iter()
-            .map(parse_goal_player_context)
-            .collect::<SubtrActorResult<Vec<_>>>()?,
-    })
-}
-
-fn parse_goal_player_context(value: &Value) -> SubtrActorResult<GoalPlayerContext> {
-    let object = json_object(value, "goal player context")?;
-    Ok(GoalPlayerContext {
-        player: json_required_remote_id(object, "player")?,
-        is_team_0: json_required_bool(object, "is_team_0")?,
-        position: json_optional_goal_context_position(object.get("position"))?,
-        boost_amount: json_optional_f32(object.get("boost_amount"))?,
-        average_boost_in_leadup: json_optional_f32(object.get("average_boost_in_leadup"))?,
-        min_boost_in_leadup: json_optional_f32(object.get("min_boost_in_leadup"))?,
-        is_most_back: json_required_bool(object, "is_most_back")?,
-    })
-}
-
-fn parse_goal_touch_context(value: &Value) -> SubtrActorResult<GoalTouchContext> {
-    let object = json_object(value, "goal touch context")?;
-    Ok(GoalTouchContext {
-        time: json_required_f32(object, "time")?,
-        frame: json_required_usize(object, "frame")?,
-        player: json_required_remote_id(object, "player")?,
-        is_team_0: json_required_bool(object, "is_team_0")?,
-        ball_position: json_optional_goal_context_position(object.get("ball_position"))?,
-        ball_speed_after_touch: json_optional_f32(object.get("ball_speed_after_touch"))?,
-        player_position: json_optional_goal_context_position(object.get("player_position"))?,
-        players: match object.get("players").and_then(Value::as_array) {
-            Some(players) => players
-                .iter()
-                .map(parse_goal_player_context)
-                .collect::<SubtrActorResult<Vec<_>>>()?,
-            None => Vec::new(),
-        },
-    })
-}
-
-fn parse_backboard_event(value: &Value) -> SubtrActorResult<BackboardBounceEvent> {
-    let object = json_object(value, "backboard event")?;
-    Ok(BackboardBounceEvent {
-        time: json_required_f32(object, "time")?,
-        frame: json_required_usize(object, "frame")?,
-        player: json_required_remote_id(object, "player")?,
-        player_position: json_optional_vec3(object.get("player_position"))?,
-        is_team_0: json_required_bool(object, "is_team_0")?,
-    })
-}
-
-fn parse_ceiling_shot_event(value: &Value) -> SubtrActorResult<CeilingShotEvent> {
-    let object = json_object(value, "ceiling shot event")?;
-    Ok(CeilingShotEvent {
-        time: json_required_f32(object, "time")?,
-        frame: json_required_usize(object, "frame")?,
-        player: json_required_remote_id(object, "player")?,
-        player_position: json_optional_vec3(object.get("player_position"))?,
-        is_team_0: json_required_bool(object, "is_team_0")?,
-        ceiling_contact_time: json_required_f32(object, "ceiling_contact_time")?,
-        ceiling_contact_frame: json_required_usize(object, "ceiling_contact_frame")?,
-        time_since_ceiling_contact: json_required_f32(object, "time_since_ceiling_contact")?,
-        ceiling_contact_position: json_required_vec3(object, "ceiling_contact_position")?,
-        touch_position: json_required_vec3(object, "touch_position")?,
-        local_ball_position: json_required_vec3(object, "local_ball_position")?,
-        separation_from_ceiling: json_required_f32(object, "separation_from_ceiling")?,
-        roof_alignment: json_required_f32(object, "roof_alignment")?,
-        forward_alignment: json_required_f32(object, "forward_alignment")?,
-        forward_approach_speed: json_required_f32(object, "forward_approach_speed")?,
-        ball_speed_change: json_required_f32(object, "ball_speed_change")?,
-        confidence: json_required_f32(object, "confidence")?,
-    })
-}
-
-fn parse_wall_aerial_event(value: &Value) -> SubtrActorResult<WallAerialEvent> {
-    let object = json_object(value, "wall aerial event")?;
-    let time = json_required_f32(object, "time")?;
-    let frame = json_required_usize(object, "frame")?;
-    Ok(WallAerialEvent {
-        time,
-        frame,
-        sample_time: json_optional_f32(object.get("sample_time"))?.unwrap_or(time),
-        sample_frame: json_optional_usize(object.get("sample_frame"))?.unwrap_or(frame),
-        player: json_required_remote_id(object, "player")?,
-        is_team_0: json_required_bool(object, "is_team_0")?,
-        wall: decode_json_value(json_required_value(object, "wall")?.clone())?,
-        wall_contact_time: json_required_f32(object, "wall_contact_time")?,
-        wall_contact_frame: json_required_usize(object, "wall_contact_frame")?,
-        takeoff_time: json_required_f32(object, "takeoff_time")?,
-        takeoff_frame: json_required_usize(object, "takeoff_frame")?,
-        time_since_takeoff: json_required_f32(object, "time_since_takeoff")?,
-        wall_contact_position: json_required_vec3(object, "wall_contact_position")?,
-        takeoff_position: json_required_vec3(object, "takeoff_position")?,
-        player_position: json_required_vec3(object, "player_position")?,
-        ball_position: json_required_vec3(object, "ball_position")?,
-        setup_start_time: json_required_f32(object, "setup_start_time")?,
-        setup_start_frame: json_required_usize(object, "setup_start_frame")?,
-        setup_duration: json_required_f32(object, "setup_duration")?,
-        ball_speed: json_required_f32(object, "ball_speed")?,
-        ball_speed_change: json_required_f32(object, "ball_speed_change")?,
-        goal_alignment: json_required_f32(object, "goal_alignment")?,
-        confidence: json_required_f32(object, "confidence")?,
-    })
-}
-
-fn parse_wall_aerial_shot_event(value: &Value) -> SubtrActorResult<WallAerialShotEvent> {
-    let object = json_object(value, "wall aerial shot event")?;
-    Ok(WallAerialShotEvent {
-        time: json_required_f32(object, "time")?,
-        frame: json_required_usize(object, "frame")?,
-        player: json_required_remote_id(object, "player")?,
-        is_team_0: json_required_bool(object, "is_team_0")?,
-        wall: decode_json_value(json_required_value(object, "wall")?.clone())?,
-        wall_contact_time: json_required_f32(object, "wall_contact_time")?,
-        wall_contact_frame: json_required_usize(object, "wall_contact_frame")?,
-        takeoff_time: json_required_f32(object, "takeoff_time")?,
-        takeoff_frame: json_required_usize(object, "takeoff_frame")?,
-        time_since_takeoff: json_required_f32(object, "time_since_takeoff")?,
-        wall_contact_position: json_required_vec3(object, "wall_contact_position")?,
-        takeoff_position: json_required_vec3(object, "takeoff_position")?,
-        player_position: json_required_vec3(object, "player_position")?,
-        ball_position: json_required_vec3(object, "ball_position")?,
-        ball_speed: json_optional_f32(object.get("ball_speed"))?,
-        goal_alignment: json_optional_f32(object.get("goal_alignment"))?,
-        confidence: json_required_f32(object, "confidence")?,
-    })
-}
-
-fn parse_center_event(value: &Value) -> SubtrActorResult<CenterEvent> {
-    let object = json_object(value, "center event")?;
-    Ok(CenterEvent {
-        time: json_required_f32(object, "time")?,
-        frame: json_required_usize(object, "frame")?,
-        player: json_required_remote_id(object, "player")?,
-        player_position: json_optional_vec3(object.get("player_position"))?,
-        is_team_0: json_required_bool(object, "is_team_0")?,
-        start_time: json_required_f32(object, "start_time")?,
-        start_frame: json_required_usize(object, "start_frame")?,
-        duration: json_required_f32(object, "duration")?,
-        start_ball_position: json_required_vec3(object, "start_ball_position")?,
-        end_ball_position: json_required_vec3(object, "end_ball_position")?,
-        ball_travel_distance: json_required_f32(object, "ball_travel_distance")?,
-        ball_advance_distance: json_required_f32(object, "ball_advance_distance")?,
-        lateral_centering_distance: json_required_f32(object, "lateral_centering_distance")?,
-    })
-}
-
-fn parse_double_tap_event(value: &Value) -> SubtrActorResult<DoubleTapEvent> {
-    let object = json_object(value, "double tap event")?;
-    Ok(DoubleTapEvent {
-        time: json_required_f32(object, "time")?,
-        frame: json_required_usize(object, "frame")?,
-        player: json_required_remote_id(object, "player")?,
-        player_position: json_optional_vec3(object.get("player_position"))?,
-        is_team_0: json_required_bool(object, "is_team_0")?,
-        backboard_time: json_required_f32(object, "backboard_time")?,
-        backboard_frame: json_required_usize(object, "backboard_frame")?,
-    })
-}
-
-fn parse_pass_event(value: &Value) -> SubtrActorResult<PassEvent> {
-    let object = json_object(value, "pass event")?;
-    let time = json_required_f32(object, "time")?;
-    let frame = json_required_usize(object, "frame")?;
-    Ok(PassEvent {
-        time,
-        frame,
-        sample_time: json_optional_f32(object.get("sample_time"))?.unwrap_or(time),
-        sample_frame: json_optional_usize(object.get("sample_frame"))?.unwrap_or(frame),
-        passer: json_required_remote_id(object, "passer")?,
-        passer_position: json_optional_vec3(object.get("passer_position"))?,
-        receiver: json_required_remote_id(object, "receiver")?,
-        receiver_position: json_optional_vec3(object.get("receiver_position"))?,
-        is_team_0: json_required_bool(object, "is_team_0")?,
-        start_time: json_required_f32(object, "start_time")?,
-        start_frame: json_required_usize(object, "start_frame")?,
-        duration: json_required_f32(object, "duration")?,
-        ball_travel_distance: json_required_f32(object, "ball_travel_distance")?,
-        ball_advance_distance: json_required_f32(object, "ball_advance_distance")?,
-        pass_kind: parse_pass_kind(object.get("pass_kind"))?,
-    })
-}
-
-fn parse_pass_last_completed_event(value: &Value) -> SubtrActorResult<PassLastCompletedEvent> {
-    let object = json_object(value, "pass last completed event")?;
-    Ok(PassLastCompletedEvent {
-        time: json_required_f32(object, "time")?,
-        frame: json_required_usize(object, "frame")?,
-        player: json_optional_remote_id(object.get("player"))?,
-        player_position: json_optional_vec3(object.get("player_position"))?,
-    })
-}
-
-fn parse_pass_kind(value: Option<&Value>) -> SubtrActorResult<PassKind> {
-    let Some(value) = value else {
-        return Ok(PassKind::Direct);
-    };
-    let kind = value.as_str().ok_or_else(|| {
-        SubtrActorError::new(SubtrActorErrorVariant::StatsSerializationError(
-            "Expected JSON field 'pass_kind' to be a string".to_owned(),
-        ))
-    })?;
-    match kind {
-        "direct" => Ok(PassKind::Direct),
-        "backboard" => Ok(PassKind::Backboard),
-        "fifty_fifty" => Ok(PassKind::FiftyFifty),
-        "fifty_fifty_backboard" => Ok(PassKind::FiftyFiftyBackboard),
-        other => Err(SubtrActorError::new(
-            SubtrActorErrorVariant::StatsSerializationError(format!("Unknown pass kind '{other}'")),
-        )),
-    }
-}
-
-fn parse_ball_carry_event(value: &Value) -> SubtrActorResult<BallCarryEvent> {
-    let object = json_object(value, "ball carry event")?;
-    Ok(BallCarryEvent {
-        player_id: json_required_remote_id(object, "player_id")?,
-        start_position: json_required_vec3(object, "start_position")?,
-        end_position: json_required_vec3(object, "end_position")?,
-        is_team_0: json_required_bool(object, "is_team_0")?,
-        kind: parse_ball_carry_kind(json_required_str(object, "kind")?)?,
-        start_frame: json_required_usize(object, "start_frame")?,
-        end_frame: json_required_usize(object, "end_frame")?,
-        start_time: json_required_f32(object, "start_time")?,
-        end_time: json_required_f32(object, "end_time")?,
-        duration: json_required_f32(object, "duration")?,
-        straight_line_distance: json_required_f32(object, "straight_line_distance")?,
-        path_distance: json_required_f32(object, "path_distance")?,
-        average_horizontal_gap: json_required_f32(object, "average_horizontal_gap")?,
-        average_vertical_gap: json_required_f32(object, "average_vertical_gap")?,
-        average_speed: json_required_f32(object, "average_speed")?,
-        touch_count: json_required_usize(object, "touch_count")? as u32,
-        air_touch_count: json_required_usize(object, "air_touch_count")? as u32,
-        air_dribble_origin: parse_air_dribble_origin(object.get("air_dribble_origin"))?,
-    })
-}
-
-fn parse_ball_carry_kind(kind: &str) -> SubtrActorResult<BallCarryKind> {
-    match kind {
-        "carry" => Ok(BallCarryKind::Carry),
-        "air_dribble" => Ok(BallCarryKind::AirDribble),
-        other => Err(SubtrActorError::new(
-            SubtrActorErrorVariant::StatsSerializationError(format!(
-                "Unknown ball carry kind '{other}'"
-            )),
-        )),
-    }
-}
-
-fn parse_air_dribble_origin(value: Option<&Value>) -> SubtrActorResult<Option<AirDribbleOrigin>> {
-    let Some(value) = value else {
-        return Ok(None);
-    };
-    if value.is_null() {
-        return Ok(None);
-    }
-    let origin = value.as_str().ok_or_else(|| {
-        SubtrActorError::new(SubtrActorErrorVariant::StatsSerializationError(
-            "Expected optional JSON field 'air_dribble_origin' to be a string".to_owned(),
-        ))
-    })?;
-    match origin {
-        "ground_to_air" => Ok(Some(AirDribbleOrigin::GroundToAir)),
-        "wall_to_air" => Ok(Some(AirDribbleOrigin::WallToAir)),
-        other => Err(SubtrActorError::new(
-            SubtrActorErrorVariant::StatsSerializationError(format!(
-                "Unknown air dribble origin '{other}'"
-            )),
-        )),
-    }
-}
-
-fn parse_one_timer_event(value: &Value) -> SubtrActorResult<OneTimerEvent> {
-    let object = json_object(value, "one timer event")?;
-    Ok(OneTimerEvent {
-        time: json_required_f32(object, "time")?,
-        frame: json_required_usize(object, "frame")?,
-        player: json_required_remote_id(object, "player")?,
-        player_position: json_optional_vec3(object.get("player_position"))?,
-        passer: json_required_remote_id(object, "passer")?,
-        passer_position: json_optional_vec3(object.get("passer_position"))?,
-        is_team_0: json_required_bool(object, "is_team_0")?,
-        pass_start_time: json_required_f32(object, "pass_start_time")?,
-        pass_start_frame: json_required_usize(object, "pass_start_frame")?,
-        pass_duration: json_required_f32(object, "pass_duration")?,
-        pass_travel_distance: json_required_f32(object, "pass_travel_distance")?,
-        pass_advance_distance: json_required_f32(object, "pass_advance_distance")?,
-        ball_speed: json_required_f32(object, "ball_speed")?,
-        goal_alignment: json_required_f32(object, "goal_alignment")?,
-    })
-}
-
-fn parse_half_volley_event(value: &Value) -> SubtrActorResult<HalfVolleyEvent> {
-    let object = json_object(value, "half volley event")?;
-    let time = json_required_f32(object, "time")?;
-    let frame = json_required_usize(object, "frame")?;
-    Ok(HalfVolleyEvent {
-        time,
-        frame,
-        sample_time: json_optional_f32(object.get("sample_time"))?.unwrap_or(time),
-        sample_frame: json_optional_usize(object.get("sample_frame"))?.unwrap_or(frame),
-        player: json_required_remote_id(object, "player")?,
-        player_position: json_optional_vec3(object.get("player_position"))?,
-        is_team_0: json_required_bool(object, "is_team_0")?,
-        bounce_time: json_required_f32(object, "bounce_time")?,
-        bounce_frame: json_required_usize(object, "bounce_frame")?,
-        bounce_to_touch_seconds: json_required_f32(object, "bounce_to_touch_seconds")?,
-        ball_speed: json_required_f32(object, "ball_speed")?,
-        goal_alignment: json_required_f32(object, "goal_alignment")?,
-    })
-}
-
-fn parse_goal_tag_event(value: &Value) -> SubtrActorResult<GoalTagEvent> {
-    let object = json_object(value, "goal tag event")?;
-    Ok(GoalTagEvent {
-        goal_index: json_required_usize(object, "goal_index")?,
-        time: json_required_f32(object, "time")?,
-        frame: json_required_usize(object, "frame")?,
-        kind: decode_json_value(json_required_value(object, "kind")?.clone())?,
-        scoring_team_is_team_0: json_required_bool(object, "scoring_team_is_team_0")?,
-        scorer: json_optional_remote_id(object.get("scorer"))?,
-        scorer_position: json_optional_goal_context_position(object.get("scorer_position"))?,
-        confidence: json_required_f32(object, "confidence")?,
-        modifiers: json_optional_array(object.get("modifiers"))?
-            .iter()
-            .map(|modifier| decode_json_value(modifier.clone()))
-            .collect::<SubtrActorResult<Vec<_>>>()?,
-        evidence: json_required_array(object, "evidence")?
-            .iter()
-            .map(parse_goal_tag_evidence)
-            .collect::<SubtrActorResult<Vec<_>>>()?,
-    })
-}
-
-fn parse_goal_tag_evidence(value: &Value) -> SubtrActorResult<GoalTagEvidence> {
-    let object = json_object(value, "goal tag evidence")?;
-    Ok(GoalTagEvidence {
-        kind: decode_json_value(json_required_value(object, "kind")?.clone())?,
-        time: json_required_f32(object, "time")?,
-        frame: json_required_usize(object, "frame")?,
-        player: json_optional_remote_id(object.get("player"))?,
-        player_position: json_optional_goal_context_position(object.get("player_position"))?,
-    })
-}
-
-fn parse_fifty_fifty_event(value: &Value) -> SubtrActorResult<FiftyFiftyEvent> {
-    let object = json_object(value, "fifty fifty event")?;
-    Ok(FiftyFiftyEvent {
-        start_time: json_required_f32(object, "start_time")?,
-        start_frame: json_required_usize(object, "start_frame")?,
-        resolve_time: json_required_f32(object, "resolve_time")?,
-        resolve_frame: json_required_usize(object, "resolve_frame")?,
-        is_kickoff: json_required_bool(object, "is_kickoff")?,
-        team_zero_player: json_optional_remote_id(object.get("team_zero_player"))?,
-        team_one_player: json_optional_remote_id(object.get("team_one_player"))?,
-        team_zero_touch_time: json_optional_f32(object.get("team_zero_touch_time"))?,
-        team_zero_touch_frame: json_optional_usize(object.get("team_zero_touch_frame"))?,
-        team_zero_dodge_contact: json_optional_bool(object.get("team_zero_dodge_contact"))
-            .unwrap_or(false),
-        team_one_touch_time: json_optional_f32(object.get("team_one_touch_time"))?,
-        team_one_touch_frame: json_optional_usize(object.get("team_one_touch_frame"))?,
-        team_one_dodge_contact: json_optional_bool(object.get("team_one_dodge_contact"))
-            .unwrap_or(false),
-        team_zero_position: json_required_vec3(object, "team_zero_position")?,
-        team_one_position: json_required_vec3(object, "team_one_position")?,
-        midpoint: json_required_vec3(object, "midpoint")?,
-        plane_normal: json_required_vec3(object, "plane_normal")?,
-        winning_team_is_team_0: json_optional_bool(object.get("winning_team_is_team_0")),
-        possession_team_is_team_0: json_optional_bool(object.get("possession_team_is_team_0")),
-    })
-}
-
-fn parse_speed_flip_event(value: &Value) -> SubtrActorResult<SpeedFlipEvent> {
-    let object = json_object(value, "speed flip event")?;
-    let time = json_required_f32(object, "time")?;
-    let frame = json_required_usize(object, "frame")?;
-    Ok(SpeedFlipEvent {
-        time,
-        frame,
-        resolved_time: json_optional_f32(object.get("resolved_time"))?.unwrap_or(time),
-        resolved_frame: json_optional_usize(object.get("resolved_frame"))?.unwrap_or(frame),
-        player: json_required_remote_id(object, "player")?,
-        is_team_0: json_required_bool(object, "is_team_0")?,
-        time_since_kickoff_start: json_required_f32(object, "time_since_kickoff_start")?,
-        start_position: json_required_vec3(object, "start_position")?,
-        end_position: json_required_vec3(object, "end_position")?,
-        start_speed: json_required_f32(object, "start_speed")?,
-        max_speed: json_required_f32(object, "max_speed")?,
-        best_alignment: json_required_f32(object, "best_alignment")?,
-        diagonal_score: json_required_f32(object, "diagonal_score")?,
-        cancel_score: json_required_f32(object, "cancel_score")?,
-        speed_score: json_required_f32(object, "speed_score")?,
-        confidence: json_required_f32(object, "confidence")?,
-    })
-}
-
-fn parse_half_flip_event(value: &Value) -> SubtrActorResult<HalfFlipEvent> {
-    let object = json_object(value, "half flip event")?;
-    Ok(HalfFlipEvent {
-        time: json_required_f32(object, "time")?,
-        frame: json_required_usize(object, "frame")?,
-        player: json_required_remote_id(object, "player")?,
-        is_team_0: json_required_bool(object, "is_team_0")?,
-        start_position: json_required_vec3(object, "start_position")?,
-        end_position: json_required_vec3(object, "end_position")?,
-        start_speed: json_required_f32(object, "start_speed")?,
-        end_speed: json_required_f32(object, "end_speed")?,
-        start_backward_alignment: json_required_f32(object, "start_backward_alignment")?,
-        best_reorientation_alignment: json_required_f32(object, "best_reorientation_alignment")?,
-        best_forward_reversal: json_required_f32(object, "best_forward_reversal")?,
-        max_forward_vertical: json_required_f32(object, "max_forward_vertical")?,
-        confidence: json_required_f32(object, "confidence")?,
-    })
-}
-
-fn parse_wavedash_event(value: &Value) -> SubtrActorResult<WavedashEvent> {
-    let object = json_object(value, "wavedash event")?;
-    Ok(WavedashEvent {
-        time: json_required_f32(object, "time")?,
-        frame: json_required_usize(object, "frame")?,
-        player: json_required_remote_id(object, "player")?,
-        is_team_0: json_required_bool(object, "is_team_0")?,
-        dodge_time: json_required_f32(object, "dodge_time")?,
-        dodge_frame: json_required_usize(object, "dodge_frame")?,
-        time_since_dodge: json_required_f32(object, "time_since_dodge")?,
-        dodge_position: json_required_vec3(object, "dodge_position")?,
-        landing_position: json_required_vec3(object, "landing_position")?,
-        start_speed: json_required_f32(object, "start_speed")?,
-        landing_speed: json_required_f32(object, "landing_speed")?,
-        horizontal_speed_gain: json_required_f32(object, "horizontal_speed_gain")?,
-        landing_uprightness: json_required_f32(object, "landing_uprightness")?,
-        confidence: json_required_f32(object, "confidence")?,
-    })
-}
-
-fn parse_whiff_event(value: &Value) -> SubtrActorResult<WhiffEvent> {
-    let object = json_object(value, "whiff event")?;
-    let time = json_required_f32(object, "time")?;
-    let frame = json_required_usize(object, "frame")?;
-    Ok(WhiffEvent {
-        kind: match object.get("kind").and_then(Value::as_str) {
-            None | Some("whiff") => WhiffEventKind::Whiff,
-            Some("beaten_to_ball") => WhiffEventKind::BeatenToBall,
-            Some(kind) => {
-                return SubtrActorError::new_result(
-                    SubtrActorErrorVariant::StatsSerializationError(format!(
-                        "Unknown whiff event kind '{kind}'"
-                    )),
-                );
-            }
-        },
-        time,
-        frame,
-        resolved_time: json_optional_f32(object.get("resolved_time"))?.unwrap_or(time),
-        resolved_frame: json_optional_usize(object.get("resolved_frame"))?.unwrap_or(frame),
-        player: json_required_remote_id(object, "player")?,
-        player_position: json_optional_vec3(object.get("player_position"))?,
-        is_team_0: json_required_bool(object, "is_team_0")?,
-        closest_approach_distance: json_required_f32(object, "closest_approach_distance")?,
-        forward_alignment: json_required_f32(object, "forward_alignment")?,
-        approach_speed: json_required_f32(object, "approach_speed")?,
-        dodge_active: json_required_bool(object, "dodge_active")?,
-        aerial: json_required_bool(object, "aerial")?,
-    })
-}
-
-fn parse_bump_event(value: &Value) -> SubtrActorResult<BumpEvent> {
-    let object = json_object(value, "bump event")?;
-    Ok(BumpEvent {
-        time: json_required_f32(object, "time")?,
-        frame: json_required_usize(object, "frame")?,
-        initiator: json_required_remote_id(object, "initiator")?,
-        victim: json_required_remote_id(object, "victim")?,
-        initiator_is_team_0: json_required_bool(object, "initiator_is_team_0")?,
-        victim_is_team_0: json_required_bool(object, "victim_is_team_0")?,
-        is_team_bump: json_required_bool(object, "is_team_bump")?,
-        strength: json_required_f32(object, "strength")?,
-        confidence: json_required_f32(object, "confidence")?,
-        contact_distance: json_required_f32(object, "contact_distance")?,
-        closing_speed: json_required_f32(object, "closing_speed")?,
-        victim_impulse: json_required_f32(object, "victim_impulse")?,
-        initiator_position: json_required_vec3(object, "initiator_position")?,
-        victim_position: json_required_vec3(object, "victim_position")?,
-    })
-}
-
-fn parse_boost_pickup_comparison_event(
-    value: &Value,
-) -> SubtrActorResult<BoostPickupComparisonEvent> {
-    let object = json_object(value, "boost pickup comparison event")?;
-    Ok(BoostPickupComparisonEvent {
-        comparison: decode_json_value(json_required_value(object, "comparison")?.clone())?,
-        frame: json_required_usize(object, "frame")?,
-        time: json_required_f32(object, "time")?,
-        player_id: json_required_remote_id(object, "player_id")?,
-        player_position: json_optional_vec3(object.get("player_position"))?,
-        is_team_0: json_required_bool(object, "is_team_0")?,
-        pad_type: decode_json_value(json_required_value(object, "pad_type")?.clone())?,
-        field_half: decode_json_value(json_required_value(object, "field_half")?.clone())?,
-        activity: decode_json_value(json_required_value(object, "activity")?.clone())?,
-        reported_frame: json_optional_usize(object.get("reported_frame"))?,
-        reported_time: json_optional_f32(object.get("reported_time"))?,
-        inferred_frame: json_optional_usize(object.get("inferred_frame"))?,
-        inferred_time: json_optional_f32(object.get("inferred_time"))?,
-        boost_before: json_optional_f32(object.get("boost_before"))?,
-        boost_after: json_optional_f32(object.get("boost_after"))?,
-    })
-}
-
-fn parse_boost_ledger_event(value: &Value) -> SubtrActorResult<BoostLedgerEvent> {
-    let object = json_object(value, "boost ledger event")?;
-    Ok(BoostLedgerEvent {
-        frame: json_required_usize(object, "frame")?,
-        time: json_required_f32(object, "time")?,
-        player_id: json_required_remote_id(object, "player_id")?,
-        player_position: json_optional_vec3(object.get("player_position"))?,
-        is_team_0: json_required_bool(object, "is_team_0")?,
-        transaction: decode_json_value(json_required_value(object, "transaction")?.clone())?,
-        amount: json_required_f32(object, "amount")?,
-        count: json_required_usize(object, "count")? as u32,
-        labels: decode_json_value(
-            object
-                .get("labels")
-                .cloned()
-                .unwrap_or_else(|| Value::Array(Vec::new())),
-        )?,
-        boost_before: json_optional_f32(object.get("boost_before"))?,
-        boost_after: json_optional_f32(object.get("boost_after"))?,
-    })
-}
-
-fn parse_boost_state_event(value: &Value) -> SubtrActorResult<BoostStateEvent> {
-    let object = json_object(value, "boost state event")?;
-    Ok(BoostStateEvent {
-        frame: json_required_usize(object, "frame")?,
-        time: json_required_f32(object, "time")?,
-        player_id: json_required_remote_id(object, "player_id")?,
-        player_position: json_optional_vec3(object.get("player_position"))?,
-        is_team_0: json_required_bool(object, "is_team_0")?,
-        boost_amount: json_required_f32(object, "boost_amount")?,
-        boost_before: json_optional_f32(object.get("boost_before"))?,
-    })
-}
-
-fn json_object<'a>(
-    value: &'a Value,
-    context: &str,
-) -> SubtrActorResult<&'a serde_json::Map<String, Value>> {
-    value.as_object().ok_or_else(|| {
-        SubtrActorError::new(SubtrActorErrorVariant::StatsSerializationError(format!(
-            "Expected {context} to be a JSON object"
-        )))
-    })
-}
-
-fn json_required_value<'a>(
-    object: &'a serde_json::Map<String, Value>,
-    field: &str,
-) -> SubtrActorResult<&'a Value> {
-    object.get(field).ok_or_else(|| {
-        SubtrActorError::new(SubtrActorErrorVariant::StatsSerializationError(format!(
-            "Missing JSON field '{field}'"
-        )))
-    })
-}
-
-fn json_required_array<'a>(
-    object: &'a serde_json::Map<String, Value>,
-    field: &str,
-) -> SubtrActorResult<&'a Vec<Value>> {
-    json_required_value(object, field)?
-        .as_array()
-        .ok_or_else(|| {
-            SubtrActorError::new(SubtrActorErrorVariant::StatsSerializationError(format!(
-                "Expected JSON field '{field}' to be an array"
-            )))
-        })
-}
-
-fn json_optional_array(value: Option<&Value>) -> SubtrActorResult<&[Value]> {
-    match value {
-        Some(Value::Array(values)) => Ok(values),
-        Some(_) => SubtrActorError::new_result(SubtrActorErrorVariant::StatsSerializationError(
-            "Expected optional JSON value to be an array".to_owned(),
-        )),
-        None => Ok(&[]),
-    }
-}
-
-fn json_f32(value: &Value) -> Option<f32> {
-    value.as_f64().map(|number| number as f32)
-}
-
-fn json_config_f32(
-    config: Option<&Map<String, Value>>,
-    key: &str,
-    legacy_key: &str,
-) -> Option<f32> {
-    config.and_then(|config| {
-        config
-            .get(key)
-            .or_else(|| config.get(legacy_key))
-            .and_then(json_f32)
-    })
-}
-
-fn json_required_f32(
-    object: &serde_json::Map<String, Value>,
-    field: &str,
-) -> SubtrActorResult<f32> {
-    json_f32(json_required_value(object, field)?).ok_or_else(|| {
-        SubtrActorError::new(SubtrActorErrorVariant::StatsSerializationError(format!(
-            "Expected JSON field '{field}' to be a float"
-        )))
-    })
-}
-
-fn json_required_usize(
-    object: &serde_json::Map<String, Value>,
-    field: &str,
-) -> SubtrActorResult<usize> {
-    json_required_value(object, field)?
-        .as_u64()
-        .map(|number| number as usize)
-        .ok_or_else(|| {
-            SubtrActorError::new(SubtrActorErrorVariant::StatsSerializationError(format!(
-                "Expected JSON field '{field}' to be an unsigned integer"
-            )))
-        })
-}
-
-fn json_required_i32(
-    object: &serde_json::Map<String, Value>,
-    field: &str,
-) -> SubtrActorResult<i32> {
-    json_required_value(object, field)?
-        .as_i64()
-        .map(|number| number as i32)
-        .ok_or_else(|| {
-            SubtrActorError::new(SubtrActorErrorVariant::StatsSerializationError(format!(
-                "Expected JSON field '{field}' to be a signed integer"
-            )))
-        })
-}
-
-fn json_required_bool(
-    object: &serde_json::Map<String, Value>,
-    field: &str,
-) -> SubtrActorResult<bool> {
-    json_required_value(object, field)?
-        .as_bool()
-        .ok_or_else(|| {
-            SubtrActorError::new(SubtrActorErrorVariant::StatsSerializationError(format!(
-                "Expected JSON field '{field}' to be a bool"
-            )))
-        })
-}
-
-fn json_required_str<'a>(
-    object: &'a serde_json::Map<String, Value>,
-    field: &str,
-) -> SubtrActorResult<&'a str> {
-    json_required_value(object, field)?.as_str().ok_or_else(|| {
-        SubtrActorError::new(SubtrActorErrorVariant::StatsSerializationError(format!(
-            "Expected JSON field '{field}' to be a string"
-        )))
-    })
-}
-
-fn json_optional_bool(value: Option<&Value>) -> Option<bool> {
-    value.and_then(Value::as_bool)
-}
-
-fn json_optional_f32(value: Option<&Value>) -> SubtrActorResult<Option<f32>> {
-    match value {
-        None | Some(Value::Null) => Ok(None),
-        Some(value) => json_f32(value).map(Some).ok_or_else(|| {
-            SubtrActorError::new(SubtrActorErrorVariant::StatsSerializationError(
-                "Expected optional JSON value to be a float".to_owned(),
-            ))
-        }),
-    }
-}
-
-fn json_optional_usize(value: Option<&Value>) -> SubtrActorResult<Option<usize>> {
-    match value {
-        None | Some(Value::Null) => Ok(None),
-        Some(value) => value
-            .as_u64()
-            .map(|number| Some(number as usize))
-            .ok_or_else(|| {
-                SubtrActorError::new(SubtrActorErrorVariant::StatsSerializationError(
-                    "Expected optional JSON value to be an unsigned integer".to_owned(),
-                ))
-            }),
-    }
-}
-
-fn json_goal_context_position(value: &Value) -> SubtrActorResult<GoalContextPosition> {
-    let object = json_object(value, "goal context position")?;
-    Ok(GoalContextPosition {
-        x: json_required_f32(object, "x")?,
-        y: json_required_f32(object, "y")?,
-        z: json_required_f32(object, "z")?,
-    })
-}
-
-fn json_optional_goal_context_position(
-    value: Option<&Value>,
-) -> SubtrActorResult<Option<GoalContextPosition>> {
-    match value {
-        None | Some(Value::Null) => Ok(None),
-        Some(value) => json_goal_context_position(value).map(Some),
-    }
-}
-
-fn json_required_vec3(
-    object: &serde_json::Map<String, Value>,
-    field: &str,
-) -> SubtrActorResult<[f32; 3]> {
-    let array = json_required_value(object, field)?
-        .as_array()
-        .ok_or_else(|| {
-            SubtrActorError::new(SubtrActorErrorVariant::StatsSerializationError(format!(
-                "Expected JSON field '{field}' to be a 3-element array"
-            )))
-        })?;
-    if array.len() != 3 {
-        return SubtrActorError::new_result(SubtrActorErrorVariant::StatsSerializationError(
-            format!("Expected JSON field '{field}' to contain exactly 3 elements"),
-        ));
-    }
-    Ok([
-        json_f32(&array[0]).ok_or_else(|| {
-            SubtrActorError::new(SubtrActorErrorVariant::StatsSerializationError(format!(
-                "Expected JSON field '{field}[0]' to be a float"
-            )))
-        })?,
-        json_f32(&array[1]).ok_or_else(|| {
-            SubtrActorError::new(SubtrActorErrorVariant::StatsSerializationError(format!(
-                "Expected JSON field '{field}[1]' to be a float"
-            )))
-        })?,
-        json_f32(&array[2]).ok_or_else(|| {
-            SubtrActorError::new(SubtrActorErrorVariant::StatsSerializationError(format!(
-                "Expected JSON field '{field}[2]' to be a float"
-            )))
-        })?,
-    ])
-}
-
-fn json_optional_vec3(value: Option<&Value>) -> SubtrActorResult<Option<[f32; 3]>> {
-    match value {
-        None | Some(Value::Null) => Ok(None),
-        Some(value) => {
-            let array = value.as_array().ok_or_else(|| {
-                SubtrActorError::new(SubtrActorErrorVariant::StatsSerializationError(
-                    "Expected optional JSON value to be a 3-element array".to_owned(),
-                ))
-            })?;
-            if array.len() != 3 {
-                return SubtrActorError::new_result(
-                    SubtrActorErrorVariant::StatsSerializationError(
-                        "Expected optional JSON value to contain exactly 3 elements".to_owned(),
-                    ),
-                );
-            }
-            Ok(Some([
-                json_f32(&array[0]).ok_or_else(|| {
-                    SubtrActorError::new(SubtrActorErrorVariant::StatsSerializationError(
-                        "Expected optional JSON value[0] to be a float".to_owned(),
-                    ))
-                })?,
-                json_f32(&array[1]).ok_or_else(|| {
-                    SubtrActorError::new(SubtrActorErrorVariant::StatsSerializationError(
-                        "Expected optional JSON value[1] to be a float".to_owned(),
-                    ))
-                })?,
-                json_f32(&array[2]).ok_or_else(|| {
-                    SubtrActorError::new(SubtrActorErrorVariant::StatsSerializationError(
-                        "Expected optional JSON value[2] to be a float".to_owned(),
-                    ))
-                })?,
-            ]))
-        }
-    }
-}
-
-fn json_required_remote_id(
-    object: &serde_json::Map<String, Value>,
-    field: &str,
-) -> SubtrActorResult<PlayerId> {
-    json_remote_id(json_required_value(object, field)?)
-}
-
-fn json_optional_remote_id(value: Option<&Value>) -> SubtrActorResult<Option<PlayerId>> {
-    match value {
-        None | Some(Value::Null) => Ok(None),
-        Some(value) => Ok(Some(json_remote_id(value)?)),
-    }
-}
-
-fn json_remote_id(value: &Value) -> SubtrActorResult<PlayerId> {
-    let object = json_object(value, "remote id")?;
-    if object.len() != 1 {
-        return SubtrActorError::new_result(SubtrActorErrorVariant::StatsSerializationError(
-            "Expected remote id to contain exactly one variant".to_owned(),
-        ));
-    }
-
-    let (variant, payload) = object.iter().next().expect("validated single variant");
-    match variant.as_str() {
-        "PlayStation" => {
-            let payload = json_object(payload, "playstation remote id")?;
-            Ok(RemoteId::PlayStation(Ps4Id {
-                online_id: json_u64(json_required_value(payload, "online_id")?)?,
-                name: json_required_value(payload, "name")?
-                    .as_str()
-                    .ok_or_else(|| {
-                        SubtrActorError::new(SubtrActorErrorVariant::StatsSerializationError(
-                            "Expected PlayStation name to be a string".to_owned(),
-                        ))
-                    })?
-                    .to_owned(),
-                unknown1: json_u8_vec(json_required_value(payload, "unknown1")?)?,
-            }))
-        }
-        "PsyNet" => {
-            let payload = json_object(payload, "psynet remote id")?;
-            Ok(RemoteId::PsyNet(PsyNetId {
-                online_id: json_u64(json_required_value(payload, "online_id")?)?,
-                unknown1: json_u8_vec(json_required_value(payload, "unknown1")?)?,
-            }))
-        }
-        "SplitScreen" => Ok(RemoteId::SplitScreen(json_u64(payload)? as u32)),
-        "Steam" => Ok(RemoteId::Steam(json_u64(payload)?)),
-        "Switch" => {
-            let payload = json_object(payload, "switch remote id")?;
-            Ok(RemoteId::Switch(SwitchId {
-                online_id: json_u64(json_required_value(payload, "online_id")?)?,
-                unknown1: json_u8_vec(json_required_value(payload, "unknown1")?)?,
-            }))
-        }
-        "Xbox" => Ok(RemoteId::Xbox(json_u64(payload)?)),
-        "QQ" => Ok(RemoteId::QQ(json_u64(payload)?)),
-        "Epic" => Ok(RemoteId::Epic(
-            payload
-                .as_str()
-                .ok_or_else(|| {
-                    SubtrActorError::new(SubtrActorErrorVariant::StatsSerializationError(
-                        "Expected Epic remote id payload to be a string".to_owned(),
-                    ))
-                })?
-                .to_owned(),
-        )),
-        variant => SubtrActorError::new_result(SubtrActorErrorVariant::StatsSerializationError(
-            format!("Unknown remote id variant '{variant}'"),
-        )),
-    }
-}
-
-fn json_u64(value: &Value) -> SubtrActorResult<u64> {
-    value
-        .as_u64()
-        .or_else(|| value.as_str().and_then(|text| text.parse().ok()))
-        .ok_or_else(|| {
-            SubtrActorError::new(SubtrActorErrorVariant::StatsSerializationError(
-                "Expected JSON value to be a u64".to_owned(),
-            ))
-        })
-}
-
-fn json_u8_vec(value: &Value) -> SubtrActorResult<Vec<u8>> {
-    value
-        .as_array()
-        .ok_or_else(|| {
-            SubtrActorError::new(SubtrActorErrorVariant::StatsSerializationError(
-                "Expected JSON value to be an array of bytes".to_owned(),
-            ))
-        })?
-        .iter()
-        .map(|entry| {
-            entry
-                .as_u64()
-                .and_then(|number| u8::try_from(number).ok())
-                .ok_or_else(|| {
-                    SubtrActorError::new(SubtrActorErrorVariant::StatsSerializationError(
-                        "Expected JSON array entry to be a byte".to_owned(),
-                    ))
-                })
-        })
-        .collect()
 }
